2021.9.5

+++ TimeBlocks/Assets/Scripts/StatsCanvas/StatsManager.cs
@@ -30,7 +30,7 @@ public class StatsManager : MonoBehaviour
         LoadOCT(dataManager.OCT);
 }
 
-    private void LoadOCT(List<int> OCT)
+    private void LoadOCT(List<double> OCT)
     {
         OCTChart.ClearData();
         OCTChart.AddXAxisData("Total");
